+++ crates/jsonschema/src/retriever.rs
@@ -60,7 +60,8 @@ impl Retrieve for DefaultRetriever {
 }
 
 #[cfg(feature = "resolve-async")]
-#[async_trait::async_trait]
+#[cfg_attr(target_family = "wasm", async_trait::async_trait(?Send))]
+#[cfg_attr(not(target_family = "wasm"), async_trait::async_trait)]
 impl referencing::AsyncRetrieve for DefaultRetriever {
     async fn retrieve(
         &self,